MatrixPilot 2.5 branch: Update to match trunk
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_2_5@484 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ altitudeCntrl.c
@@ -155,7 +155,7 @@ void normalAltitudeCntrl(void)
 		}
 		else if ( flags._.GPS_steering && desired_behavior._.land )
 		{
-			pitchAltitudeAdjust = (long) (PITCHATLAND) ;
+			pitchAltitudeAdjust = (int) (PITCHATLAND) ;
 			
 			throttleFiltered.WW += (((long)(pwTrim[THROTTLE_INPUT_CHANNEL] - throttleFiltered._.W1 ))<<THROTTLEFILTSHIFT ) ;
 			altitude_control = throttleFiltered._.W1 - throttleIn ;
